+++ core/datasource/src/main/java/com/core/datasource/model/Tag.kt
@@ -6,11 +6,11 @@ import com.core.database.model.TagEntity
  * datasource tag 정보
  */
 data class Tag(
-    val id: Int?, // tag id
+    val id: Long?, // tag id
     val name: String // tag 내용
 )
 
-internal fun Tag.toTagEntity(postId: Int) = TagEntity(
+internal fun Tag.toTagEntity(postId: Long) = TagEntity(
     id = id,
     postId = postId,
     name = name
